% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{trials}

\item{k}{successes}
}
\value{
a number or a vector of numbers
}
\description{
calculates the number of combinations
}
\examples{
bin_choose(5, 2)
5 
}
