% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{a number}

\item{prob}{a number}
}
\value{
a dataframe
}
\description{
calculates the cumluative probability
}
\examples{
bin_cumulative(1, 0.7)
success  probability cumulative
0               0.3         0.3
1               0.7         1.0
bin_cumulative(1, 0.2)
success  probability cumulative
0               0.8         0.8
1               0.2         1.0
}
