% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{a number}

\item{prob}{a number}
}
\value{
a dataframe
}
\description{
calculates the probability with all possible success
}
\examples{
bin_distribution(1, 0.7)
bin_cumulative(1, 0.2)
}
