% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{the probability to success occurs in a trial}
}
\value{
the kurtosis value
}
\description{
compute the kurtosis for distribution
}
\examples{
bin_kurtosis(5, 0.5)
-0.4
}
