% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{prob the probability to success occurs in a trial}
}
\value{
the mean value
}
\description{
compute the mean for distribution
}
\examples{
bin_choose(5, 0.5)
2.5
}
