% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{the probability to success occurs in a trial}
}
\value{
the mode value
}
\description{
compute the mode for distribution
}
\examples{
bin_mode(5, 0.5)
3
}
