% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{the probability to success occurs in a trial}
}
\value{
the skewness value
}
\description{
compute the skewness for distribution
}
\examples{
bin_skewness(5, 0.5)
0
}
