% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{a number}

\item{prob}{a number}
}
\value{
a list
}
\description{
put variables in a list
}
\examples{
bin_variable(3, 0.2)
bin_variable(3, 0.2)
}
