% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{prob the probability to success occurs in a trial}
}
\value{
the variance value
}
\description{
compute the variance for distribution
}
\examples{
bin_variance(5, 0.5)
1.25
}
