% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{a number}

\item{prob}{a number}
}
\value{
a dataframe
}
\description{
calculates the cumluative probability
}
\examples{
bin_cumulative(1, 0.7)
bin_cumulative(1, 0.2)
}
