% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{a number or a vector}

\item{trials}{a number}

\item{prob}{a number, the probability}
}
\value{
a number
}
\description{
calculates the probability
}
\examples{
bin_probability(2, 5, 0.5)
bin_probability(0:2, 5, 0.5)
}
